refactor: 디렉터리 변경 /feat: AOP
