+++ src/main/java/com/example/taskflow/log/controller/ActivityLogController.java
@@ -1,7 +1,8 @@
-package com.example.taskflow.log;
+package com.example.taskflow.log.controller;
 
 import com.example.taskflow.common.ApiResponse;
-import com.example.taskflow.log.dto.ActivityUpdateRequestDto;
+import com.example.taskflow.log.service.ActivityService;
+import com.example.taskflow.log.dto.request.ActivityUpdateRequestDto;
 import com.example.taskflow.log.dto.request.ActivityLogCreateRequestDto;
 import com.example.taskflow.log.dto.request.ActivityLogRequestDto;
 import com.example.taskflow.log.dto.response.ActivityLogResponseDto;
