+++ src/main/java/com/example/taskflow/log/dto/request/ActivityLogCreateRequestDto.java
@@ -1,8 +1,7 @@
 package com.example.taskflow.log.dto.request;
 
 
-import com.example.taskflow.log.ActivityType;
-import jakarta.validation.constraints.NotNull;
+import com.example.taskflow.log.entity.ActivityType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
