+++ src/main/java/com/example/taskflow/log/dto/response/ActivityCreateResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.taskflow.log.dto.response;
 
-import com.example.taskflow.log.ActivityType;
+import com.example.taskflow.log.entity.ActivityType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
